Add files via upload
Refactored a good portion of the code. Eslint wanted me to make a lot of global variables that I'm not sure necessarily needed to be created due to the way I'm using them (dist and cardPos for example). Let me know if you want me to create those global variables and I will. Found a bunch of repeated code that didn't need to exist that I got rid of.